Added updateDb method with custom data URL option.
- Extracted updatedb script to a updateDb() method.
- Added ability to specify a custom repository URL for data files.
- Updated tests to match latest data files.
