+++ lib/geoip.js
@@ -6,6 +6,7 @@ fs.existsSync = fs.existsSync || path.existsSync;
 
 var utils = require('./utils');
 var fsWatcher = require('./fsWatcher');
+var updateDb = require('./updateDb')
 var async = require('async');
 
 var watcherName = 'dataWatcher';
@@ -484,6 +485,8 @@ module.exports = {
 
 		return n;
 	},
+
+	updateDb: updateDb,
 
 	// Start watching for data updates. The watcher waits one minute for file transfer to 
 	// completete before triggering the callback.
